+++ src/ResourceManager/Sql/Commands.Sql/Database/Model/AzureSqlDatabaseModel.cs
@@ -24,6 +24,11 @@ namespace Microsoft.Azure.Commands.Sql.Database.Model
     public class AzureSqlDatabaseModel
     {
         /// <summary>
+        /// Template to generate database id
+        /// </summary>
+        internal const string IdTemplate = "/subscriptions/{0}/resourceGroups/{1}/providers/Microsoft.Sql/servers/{2}/databases/{3}";
+
+        /// <summary>
         /// Gets or sets the name of the resource group
         /// </summary>
         public string ResourceGroupName { get; set; }